// src/types/interfaces.ts
export interface IFormFields {
  title: string;
  description: string;
  heading: string;
  // chat: string;
  // id?
}

export interface IDataFields {
  title: string | undefined;
  description: string | undefined;
  heading: string | undefined;
  // chat: string;
  id: string;
  // [x: string]: nev
}